1.12.2-3.0.5
Added code to return the actual block you placed down when it is determined that you can no longer place blocks.

Added a config to set when it is that a column is no longer a column but an ugly nerd pole. Default 10.

Thanks to Raycoms of Minecolonies for the helpful code suggestion.